:wrench: add script to setup and run backend (#145)
The script tries to determine if pg_config is available and offers to
install it via Homebrew (only on macOS, could potentially be expanded to
support apt-get, etc.).
Script then tries to determine which version of Python is the default.
If default is Python 2 and Python 3 is available, the script will try to
run commands using Python 3. If Python 3 isn't available it will offer
to install it, again via Homebrew.